refactor bridge
